count 10 sec. Let "isTimeOut" be true after 10 sec
