+++ apps/Flashcard.js
@@ -1,21 +1,11 @@
-import React, { useEffect } from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import React, { useEffect, useState } from 'react';
+import { View, Text, StyleSheet, Image } from 'react-native';
 
 import vocab from '../assets/flashcard.json';
 import AnswerChoice from '../components/flashcard/AnswerChoice';
+import CountTenSec from '../components/flashcard/CountTenSec';
 
 const Flashcard = () => {
-  const [countTenSec, setSeconds] = useState(10);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setSeconds((countTenSec) => countTenSec - 1);
-    }, 1000);
-    return () => {
-      clearInterval(interval);
-    };
-  }, []);
-
   const random_index = Math.floor(Math.random() * vocab.length);
   let answer_choices_index = [random_index];
   while (answer_choices_index.length < 4) {
@@ -25,9 +15,10 @@ const Flashcard = () => {
         continue;
       }
     }
-
     answer_choices_index.push(r);
   }
+
+  const [isTimeOut, setIsTimeOut] = useState(false);
 
   const problem = vocab[random_index];
   const problem_eng = problem.English;
@@ -37,6 +28,11 @@ const Flashcard = () => {
 
   return (
     <View style={styles.container}>
+      <Image style={styles.image} source={require('../assets/correct.jpg')} />
+      <CountTenSec isTimeOut={isTimeOut} setIsTimeOut={setIsTimeOut} />
+      <View>
+        <Text>{isTimeOut ? 'true' : 'false'}</Text>
+      </View>
       <Text style={styles.eng}>{problem_eng}</Text>
       <View style={styles.answerContainer}>
         <AnswerChoice
@@ -65,11 +61,24 @@ const Flashcard = () => {
 };
 
 const styles = StyleSheet.create({
-  container: {},
+  container: {
+    // flex: 1,
+    // flexDirection: 'column',
+    // justifyContent: 'center',
+    // alignItems: 'center',
+  },
   answerContainer: {},
   eng: {},
   word: {},
   description: {},
+  image: {
+    position: 'absolute',
+    // flexDirection: 'column',
+    // justifyContent: 'center',
+    alignContent: 'center',
+    width: 300,
+    height: 300,
+  },
 });
 
 export default Flashcard;
